requirements.txt
apache-airflow-providers-google==16.1.0
requests==2.32.4
pendulum==3.1.0
pandas==2.3.1
numpy==2.2.6
